+++ Magic/docs/release-notes.md
@@ -2,9 +2,12 @@
 
 # Release notes
 ## Upcoming release
-* Improved spell icons to better fit vanilla style (thanks to Ash!).
-* Added magic school icons.
-* Added tooltips.
+* Improved magic UI:
+  * Added icons for magic school tabs.
+  * Added tooltips.
+  * Replaced spell icons (thanks to Ash!).
+  * You can no longer select a magic school you don't know any spells for.
+  * Various UI tweaks.
 
 ## 0.6.2
 Released 29 October 2021 for SMAPI 3.12.5 or later. Updated by Pathoschild.